requirements.txt
tensorflow-gpu == 1.14
opencv-python
scipy
numpy

